refactor(ajsttil): update TypeMetadata without xor metadata
