+++ packages/transformers/json-schema-2-ajsttil/src/jsonSchema/202012/calculations/transformJsonSchema.ts
@@ -103,6 +103,10 @@ function handleJsonSchemaChildren(
           kind: TypeMetadataKind.floatType,
         },
         {
+          kind: TypeMetadataKind.literalType,
+          literal: null,
+        },
+        {
           kind: TypeMetadataKind.objectType,
         },
         {
@@ -198,7 +202,7 @@ function handleJsonSchemaSubschemas(
       children: schema.oneOf.map((schema: JsonSchema) =>
         transformJsonSchema(schema, context),
       ),
-      kind: TypeMetadataKind.xor,
+      kind: TypeMetadataKind.or,
     });
   }
 }
